+++ src/main/java/znu/visum/components/genres/domain/models/Genre.java
@@ -1,7 +1,5 @@
 package znu.visum.components.genres.domain.models;
 
-import java.util.Objects;
-
 public class Genre {
   private Long id;
 
@@ -26,18 +24,5 @@ public class Genre {
 
   public void setType(String type) {
     this.type = type;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    Genre genre = (Genre) o;
-    return Objects.equals(getId(), genre.getId()) && getType().equals(genre.getType());
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(getId(), getType());
   }
 }